Fix memory leak for disk checkpointing (#4020)
* deepcopy name and count function attributes
* Avoid reference the callable function name; Disk checkpoint using checkpoint_schedules
